% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restCalls.R
\name{getRestCall}
\alias{getRestCall}
\title{Rest call functions}
\usage{
getRestCall(baseUrl = "https://icanhazdadjoke.com", endPoint = "",
  queryParams = NULL, returnType = c("application/json", "text/plain",
  "application/json"))
}
\arguments{
\item{baseUrl}{character. Base website, shouldn't change}

\item{endPoint}{character. API endpoint}

\item{queryParams}{list. Named list of query parameters}

\item{returnType}{character. Return format}
}
\value{
httr object
}
\description{
These functions are just to make specific rest calls to I Can Haz Dad Joke.
}
