% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tell_joke.R
\name{tell_joke}
\alias{tell_joke}
\alias{print.joke}
\alias{print.jokes}
\title{Tell dad jokes}
\source{
https://www.fatherhood.gov/for-dads/dad-jokes
}
\usage{
tell_joke(n.jokes = 1L, as.character = FALSE, random = TRUE)

\method{print}{joke}(x, ...)

\method{print}{jokes}(x, ...)
}
\arguments{
\item{n.jokes}{integer specifying number of jokes to return.
Defaults to 1.}

\item{as.character}{logical indicating whether or not to return a character vector.
Defaults to \code{FALSE}}

\item{random}{logical indicating whether or not to choose \code{n.jokes} at random.
Defaults to \code{TRUE}}
}
\value{
A (printable) list of jokes (default) or a character vector (if \code{as.character = TRUE})
}
\description{
Function returns a dad joke either by printing it to the console or
    by returning a is as a character vector
}
\examples{
\dontrun{
tell_joke()
tell_joke(as.character = TRUE)
}
}
